Added event controller/view folder
