+++ Radabite/Client/WebClient/Views/Event/Event.cshtml
@@ -1,0 +1,9 @@
+﻿@{
+	ViewBag.Title = "Event";
+}
+
+<div class="row">
+	<hgroup class="title">
+		<h2>@ViewBag.Message</h2>
+	</hgroup>
+</div>
