+++ Radabite/Client/WebClient/Views/Shared/_NavigationBar.cshtml
@@ -11,9 +11,9 @@
 
         <section class="top-bar-section">
             <ul class="left">
-                <li>@Html.ActionLink("Profile", "UserProfile", new { userId = Model["userId"] })</li>
-                <li>@Html.ActionLink("Create", "CreateEvent", new { userId = Model["userId"] })</li>
-                <li>@Html.ActionLink("Discover", "DiscoverEvent", new { userId = Model["userId"] })</li>
+                <li>@Html.ActionLink("Profile", "../Event/UserProfile", new { userId = Model["userId"] })</li>
+                <li>@Html.ActionLink("Create", "../Event/CreateEvent", new { userId = Model["userId"] })</li>
+                <li>@Html.ActionLink("Discover", "../Event/DiscoverEvent", new { userId = Model["userId"] })</li>
                 
             </ul>
             <ul class="right">
