+++ Radabite/Radabite.csproj
@@ -206,6 +206,7 @@
     <Compile Include="Backend\Interfaces\IUserManager.cs" />
     <Compile Include="Backend\Managers\UserManager.cs" />
     <Compile Include="Bindings.cs" />
+    <Compile Include="Client\WebClient\Controllers\EventController.cs" />
     <Compile Include="Client\WebClient\ExtendedRazor.cs" />
     <Compile Include="Client\WebClient\App_Start\AuthConfig.cs" />
     <Compile Include="Client\WebClient\App_Start\BundleConfig.cs" />
@@ -357,9 +358,10 @@
     <Content Include="Client\WebClient\Views\Shared\_Foundation.cshtml" />
     <Content Include="Client\WebClient\Views\Home\Index.cshtml" />
     <Content Include="Client\WebClient\Views\Home\UserProfile.cshtml" />
-    <Content Include="Client\WebClient\Views\Home\CreateEvent.cshtml" />
-    <Content Include="Client\WebClient\Views\Home\DiscoverEvent.cshtml" />
+    <Content Include="Client\WebClient\Views\Event\CreateEvent.cshtml" />
+    <Content Include="Client\WebClient\Views\Event\DiscoverEvent.cshtml" />
     <Content Include="Client\WebClient\Views\Shared\_NavigationBar.cshtml" />
+    <Content Include="Client\WebClient\Views\Event\Event.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Client\WebClient\App_Data\" />
